Extract topic raw window scanner
